refactor(data.js): rename as sprites.js
