+++ src/sprites.js
